Extract update status event client into own library
