+++ services/crawl/libs/event-client-library/index.ts
@@ -1,0 +1,4 @@
+import EventClient from "./interfaces/EventClient";
+import EventBridgeClient from "./clients/EventBridgeClient";
+
+export { EventClient, EventBridgeClient };
