+++ services/crawl/libs/event-client-library/package.json
@@ -1,0 +1,8 @@
+{
+    "name": "buzzword-crawl-event-client-library",
+    "version": "1.0.0",
+    "main": "./index.js",
+    "dependencies": {
+        "@aws-sdk/client-eventbridge": "3.55.0"
+    }
+}
