+++ services/crawl/package.json
@@ -4,7 +4,8 @@
     "devDependencies": {
         "buzzword-crawl-common": "1.0.0",
         "buzzword-crawl-content-repository-library": "1.0.0",
-        "buzzword-crawl-urls-repository-library": "1.0.0"
+        "buzzword-crawl-urls-repository-library": "1.0.0",
+        "buzzword-crawl-event-client-library": "1.0.0"
     },
     "scripts": {
         "postinstall": "npx graphql-codegen --config ./codegen.yml"
